+++ Properties/AssemblyInfo.cs
@@ -21,6 +21,6 @@ using System.Windows;
     ResourceDictionaryLocation.SourceAssembly
 )]
 
-[assembly: AssemblyVersion("0.1.13")]
-[assembly: AssemblyFileVersion("0.1.13")]
-[assembly: AssemblyInformationalVersion("0.1.13")]
+[assembly: AssemblyVersion("0.1.14")]
+[assembly: AssemblyFileVersion("0.1.14")]
+[assembly: AssemblyInformationalVersion("0.1.14")]